backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
spacy==3.7.2
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
faker==20.1.0